Refactor Entity Mappings to reduce module size
