+++ custom_components/mastertherm/const.py
@@ -1,4 +1,18 @@
 """Constants used by Home Assistant Components."""
+from dataclasses import dataclass, field
+
+from homeassistant.const import Platform, UnitOfTemperature
+from homeassistant.components.climate import (
+    ClimateEntityDescription,
+    DEFAULT_MAX_TEMP,
+    DEFAULT_MIN_TEMP,
+)
+from homeassistant.components.number import NumberEntityDescription, NumberDeviceClass
+from homeassistant.components.sensor import SensorEntityDescription
+from homeassistant.components.binary_sensor import BinarySensorEntityDescription
+from homeassistant.components.select import SelectEntityDescription
+from homeassistant.components.switch import SwitchEntityDescription
+
 NAME = "Mastertherm"
 DOMAIN = "mastertherm"
 
@@ -9,3 +23,64 @@ API_VERSIONS = {
 
 # Default Refresh Rate in Seconds
 DEFAULT_REFRESH = 120
+
+
+@dataclass
+class MasterthermBinarySensorEntityDescription(BinarySensorEntityDescription):
+    """Description for the Mastertherm binary sensor entities."""
+
+
+@dataclass
+class MasterthermClimateEntityDescription(ClimateEntityDescription):
+    """Description for the Mastertherm Climate Entity."""
+
+    # Key must be the entity that decides if this is enabled
+    # in dot notation e.g. heating_circuits.hc1.ambient_requested
+    # current_temperature_path and requested_temperature_path are
+    # the entities to lookup the states in dot notation as above.
+    current_temperature_path: str = None
+    requested_temperature_path: str = None
+    min_temp: float | str = DEFAULT_MIN_TEMP
+    max_temp: float | str = DEFAULT_MAX_TEMP
+
+
+@dataclass
+class MasterthermNumberEntityDescription(NumberEntityDescription):
+    """Description for the Mastertherm Number Entity."""
+
+    device_class = NumberDeviceClass.TEMPERATURE
+    native_unit_of_measurement = UnitOfTemperature.CELSIUS
+    native_step: float = 0.1
+    mode: str = "auto"
+
+
+@dataclass
+class MasterthermSelectEntityDescription(SelectEntityDescription):
+    """Description for the Mastertherm select entities."""
+
+    options_map: dict = field(default_factory=dict)
+    read_only: bool = False
+
+
+@dataclass
+class MasterthermSensorEntityDescription(SensorEntityDescription):
+    """Description for the Mastertherm sensor entities."""
+
+    icon_state_map: dict[str, str] = field(default_factory=dict[str, str])
+
+
+@dataclass
+class MasterthermSwitchEntityDescription(SwitchEntityDescription):
+    """Description for the Mastertherm switch entities."""
+
+    read_only: bool = False
+
+
+ENTITIES: dict[str, str] = {
+    MasterthermBinarySensorEntityDescription.__name__: Platform.BINARY_SENSOR,
+    MasterthermSelectEntityDescription.__name__: Platform.SELECT,
+    MasterthermSensorEntityDescription.__name__: Platform.SENSOR,
+    MasterthermSwitchEntityDescription.__name__: Platform.SWITCH,
+    MasterthermClimateEntityDescription.__name__: Platform.CLIMATE,
+    MasterthermNumberEntityDescription.__name__: Platform.NUMBER,
+}
