+++ custom_components/mastertherm/switch.py
@@ -9,10 +9,9 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.const import CONF_ENTITIES, Platform
 
-from .const import DOMAIN
+from .const import DOMAIN, MasterthermSwitchEntityDescription
 from .coordinator import MasterthermDataUpdateCoordinator
 from .entity import MasterthermEntity
-from .entity_mappings import MasterthermSwitchEntityDescription
 
 _LOGGER = logging.getLogger(__name__)
