+++ tests/test_climate.py
@@ -11,8 +11,8 @@ from homeassistant.core import HomeAssistant
 
 from pytest_homeassistant_custom_component.common import MockConfigEntry
 
-from custom_components.mastertherm.const import DOMAIN
-from custom_components.mastertherm.entity_mappings import (
+from custom_components.mastertherm.const import (
+    DOMAIN,
     MasterthermClimateEntityDescription,
 )
