+++ tests/test_number.py
@@ -11,8 +11,8 @@ from homeassistant.components.number import (
 from homeassistant.const import Platform, ATTR_ENTITY_ID
 
 from pytest_homeassistant_custom_component.common import MockConfigEntry
-from custom_components.mastertherm.const import DOMAIN
-from custom_components.mastertherm.entity_mappings import (
+from custom_components.mastertherm.const import (
+    DOMAIN,
     MasterthermNumberEntityDescription,
 )
