+++ tests/test_select.py
@@ -13,8 +13,8 @@ from homeassistant.const import Platform, ATTR_ENTITY_ID
 from homeassistant.core import HomeAssistant
 
 from pytest_homeassistant_custom_component.common import MockConfigEntry
-from custom_components.mastertherm.const import DOMAIN
-from custom_components.mastertherm.entity_mappings import (
+from custom_components.mastertherm.const import (
+    DOMAIN,
     MasterthermSelectEntityDescription,
 )
