+++ tests/test_sensor.py
@@ -6,8 +6,8 @@ from homeassistant.core import HomeAssistant
 from homeassistant.const import Platform, UnitOfTemperature
 
 from pytest_homeassistant_custom_component.common import MockConfigEntry
-from custom_components.mastertherm.const import DOMAIN
-from custom_components.mastertherm.entity_mappings import (
+from custom_components.mastertherm.const import (
+    DOMAIN,
     MasterthermSensorEntityDescription,
 )
