+++ tests/test_switch.py
@@ -17,8 +17,8 @@ from homeassistant.const import (
 )
 
 from pytest_homeassistant_custom_component.common import MockConfigEntry
-from custom_components.mastertherm.const import DOMAIN
-from custom_components.mastertherm.entity_mappings import (
+from custom_components.mastertherm.const import (
+    DOMAIN,
     MasterthermSwitchEntityDescription,
 )
